Create mmap.asm
Linux NASM x86_64 assembly language subroutine that makes an mmap system call.

You can call this subroutine with the appropriate arguments on the stack. The return value will be in the rax register. If an error occurred, rax will be set to MAP_FAILED (-1) and errno will be set to the appropriate error code.

Note that this subroutine assumes that the calling convention is cdecl (arguments are pushed on the stack from right to left) and that the errno symbol is defined somewhere in the program's data section. You may need to modify the subroutine if your program uses a different calling convention or if errno is defined in a different location.